Conflicts:
	python/tthAnalyzeSamples_2016.py

+++ python/sbatch-node.template.sh
@@ -2,7 +2,14 @@
 # File: sbatch-node.template.sh
 # Version: 0.2
 
+# unset JAVA_HOME, because hadoop commands might not work
+# this is especially true if one has sourced necessary files for the GRID proxy
+echo 'Unsetting JAVA_HOME=$JAVA_HOME'
+unset JAVA_HOME
+
 # This value is provided by sbatchManager.py that creates sbatch scripts based this template
+echo 'Running script {{ script_file }} (created from template sbatch-node.template.sh)'
+
 
 RUNNING_COMMAND="{{ RUNNING_COMMAND }}"
 
@@ -52,10 +59,14 @@ run_wrapped_executable() {
     EXECUTABLE_LOG_FILE_NAME="`basename $EXECUTABLE_LOG_FILE`"
     TEMPORARY_EXECUTABLE_LOG_DIR="$SCRATCH_DIR/$EXECUTABLE_LOG_DIR/"
     TEMPORARY_EXECUTABLE_LOG_FILE="$TEMPORARY_EXECUTABLE_LOG_DIR/$EXECUTABLE_LOG_FILE_NAME"
+    RANDOM_SLEEP={{ random_sleep }}
 
     echo "Time is: `date`"
     echo "Hostname: `hostname`"
     echo "Current directory: `pwd`"
+
+    echo "Sleeping for $RANDOM_SLEEP seconds"
+    sleep $RANDOM_SLEEP
 
     echo "Create scratch directory: mkdir -p $SCRATCH_DIR"
     mkdir -p $SCRATCH_DIR
@@ -77,22 +88,61 @@ run_wrapped_executable() {
 
     echo "Time is: `date`"
 
-    CMSSW_SEARCH_PATH="$SCRATCH_DIR:$CMSSW_BASE/src"
-    echo "Execute command: {{ exec_name }} {{ cfg_file }} &> $TEMPORARY_EXECUTABLE_LOG_FILE"
-    {{ exec_name }} {{ cfg_file }} &> $TEMPORARY_EXECUTABLE_LOG_FILE
+    CMSSW_SEARCH_PATH="$SCRATCH_DIR:{{ cmssw_base_dir }}/src"
+
+    echo "Execute command: {{ exec_name }} {{ command_line_parameter }} &> $TEMPORARY_EXECUTABLE_LOG_FILE"
+    {{ exec_name }} {{ command_line_parameter }} &> $TEMPORARY_EXECUTABLE_LOG_FILE
     EXIT_CODE=$?
+    echo "Command {{ exec_name }} exited with code $EXIT_CODE"
 
     echo "Time is: `date`"
+
+    echo "Listing of current directory:"
+    ls -l
 
     OUTPUT_FILES="{{ outputFiles }}"
     echo "Copying output files: {{ outputFiles }}"
     for OUTPUT_FILE in $OUTPUT_FILES
     do
-      OUTPUT_FILE_SIZE=$(stat -c '%s' $OUTPUT_FILE)
-      if [ $OUTPUT_FILE_SIZE -ge 1000 ]; then
-        echo "cp $OUTPUT_FILE {{ outputDir }}"
-        cp $OUTPUT_FILE {{ outputDir }}
+      OUTPUT_DIR="{{ outputDir }}"
+      if [[ "$OUTPUT_DIR" =~ ^/hdfs* && ( ! -z $(which hadoop) ) ]]; then
+        cp_cmd="hadoop fs -copyFromLocal";
+        st_cmd="hadoop fs -stat '%b'"
+        OUTPUT_DIR=${OUTPUT_DIR#/hdfs}
       else
+        cp_cmd=cp;
+        st_cmd="stat --printf='%s'"
+      fi
+      cp_cmd="$cp_cmd -f"
+
+      OUTPUT_FILE_SIZE=$(stat -c '%s' $OUTPUT_FILE)
+      if [ -n "$OUTPUT_FILE_SIZE" ] && [ $OUTPUT_FILE_SIZE -ge 1000 ]; then
+        echo "$cp_cmd $OUTPUT_FILE $OUTPUT_DIR"
+
+        CP_RETRIES=0
+        COPIED=false
+        while [ $CP_RETRIES -lt 3 ]; do
+          CP_RETRIES=$[CP_RETRIES + 1];
+          $cp_cmd -f $OUTPUT_FILE $OUTPUT_DIR/$OUTPUT_FILE
+
+          # add a small delay before stat'ing the file
+          sleep 5s
+
+          REMOTE_SIZE=$($st_cmd $OUTPUT_DIR/$OUTPUT_FILE)
+          if [ "$REMOTE_SIZE" == "$OUTPUT_FILE_SIZE" ]; then
+            COPIED=true
+            break;
+          else
+            continue;
+          fi
+        done
+
+        if [ ! $COPIED ]; then
+          EXIT_CODE=1;
+        fi
+
+      else
+        echo "$OUTPUT_FILE is broken, will exit with code 1."
         rm $OUTPUT_FILE
         EXIT_CODE=1
       fi
@@ -110,6 +160,9 @@ run_wrapped_executable() {
     rm -r $SCRATCH_DIR
 
     echo "End time is: `date`"
+
+    echo "Waiting for 10 seconds before returning"
+    sleep 10
 
     return $EXIT_CODE
 }
